# static/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.0.4",
    "title": "Prototype API BETA",
    "description": "Demonstration of how to create a simple API and automate the deployment process",
    "termsOfService": "",
    "contact": {
      "name": "",
      "url": "https://github.com/virgesmith/protoapi",
      "email": ""
    },
    "license": {
      "name": "MIT Licence",
      "url": "https://github.com/virgesmith/protoapi/blob/master/LICENSE"
    }
  },
  "tags": [
    {
      "name": "Numeric",
      "description": "Does something numeric"
    },
    {
      "name": "Utility",
      "description": "Does something non-numeric"
    }
  ],
  "paths": {
    "/headers": {
      "get": {
        "description": "returns the client's http headers",
        "tags": [
          "Utility"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
    "/sobol": {
      "get": {
        "description": "returns a Sobol sequence",
        "tags": [
          "Numeric"
        ],
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/positive_integer"
            }
          },
          {
            "name": "length",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/positive_integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
    "/integerise": {
      "post": {
        "description": "tries to find an integer array that preserves the marginal sums",
        "tags": [
          "Numeric"
        ],
        "requestBody": {
          "description": "an n-dimensional array of floating point values with integer marginal sums",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/float_array",
                "type": "array",
                "items": {
                  "type": "number"
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "positive_integer": {
        "type": "integer",
        "minimum": 1
      },
      "float_array": {
        "description": "array of floating-point numbers, NB '3.' won't parse, use '3' or '3.0'",
        "type": "array",
        "items": {
          "type": "number"
        }
      }
    }
  }
}